requirements.txt
cycler==0.10.0
kiwisolver==1.1.0
matplotlib==3.1.1
numpy==1.16.4
Pillow==6.1.0
pyparsing==2.4.0
python-dateutil==2.8.0
scipy==1.3.0
six==1.12.0
